+++ lib/distribute_reads.rb
@@ -18,26 +18,55 @@ module DistributeReads
     lag_failover: false
   }
 
+  def self.replication_lag(connection: nil)
+    distribute_reads do
+      lag(connection: connection)
+    end
+  end
+
   def self.lag(connection: nil)
     raise DistributeReads::Error, "Don't use outside distribute_reads" unless Thread.current[:distribute_reads]
 
     begin
-      connection ||= ActiveRecord::Base.connection
-      if %w(PostgreSQL PostGIS).include?(connection.adapter_name)
-        replica_pool = connection.instance_variable_get(:@slave_pool)
-        if replica_pool && replica_pool.connections.size > 1
-          warn "[distribute_reads] Multiple replicas available, lag only reported for one"
+    connection ||= ActiveRecord::Base.connection
+
+    replica_pool = connection.instance_variable_get(:@slave_pool)
+    if replica_pool && replica_pool.connections.size > 1
+      warn "[distribute_reads] Multiple replicas available, lag only reported for one"
+    end
+
+    if %w(PostgreSQL PostGIS).include?(connection.adapter_name)
+      # cache the version number
+      @server_version_num ||= {}
+      cache_key = connection.pool.object_id
+      @server_version_num[cache_key] ||= connection.execute("SHOW server_version_num").first["server_version_num"].to_i
+
+      lag_condition =
+        if @server_version_num[cache_key] >= 100000
+          "pg_last_wal_receive_lsn() = pg_last_wal_replay_lsn()"
+        else
+          "pg_last_xlog_receive_location() = pg_last_xlog_replay_location()"
         end
 
-        connection.execute(
+      connection.execute(
         "SELECT CASE
-        WHEN NOT pg_is_in_recovery() OR pg_last_xlog_receive_location() = pg_last_xlog_replay_location() THEN 0
-        ELSE EXTRACT (EPOCH FROM NOW() - pg_last_xact_replay_timestamp())
-        END AS lag"
-        ).first["lag"].to_f
-      else
-        raise DistributeReads::Error, "Option not supported with this adapter"
+          WHEN NOT pg_is_in_recovery() OR #{lag_condition} THEN 0
+          ELSE EXTRACT (EPOCH FROM NOW() - pg_last_xact_replay_timestamp())
+        END AS lag".squish
+      ).first["lag"].to_f
+    elsif %w(MySQL Mysql2 Mysql2Spatial Mysql2Rgeo).include?(connection.adapter_name)
+      replica_value = Thread.current[:distribute_reads][:replica]
+      begin
+        # makara doesn't send SHOW queries to replica, so we must force it
+        Thread.current[:distribute_reads][:replica] = true
+        status = connection.exec_query("SHOW SLAVE STATUS").to_hash.first
+        status ? status["Seconds_Behind_Master"].to_f : 0.0
+      ensure
+        Thread.current[:distribute_reads][:replica] = replica_value
       end
+    else
+      raise DistributeReads::Error, "Option not supported with this adapter"
+    end
     rescue ActiveRecord::StatementInvalid => e
       if e.original_exception.is_a?(PG::FeatureNotSupported)
         # AWS Aurora raises the following error when checking the replication lag:
@@ -47,7 +76,7 @@ module DistributeReads
         return 0.5
       else
         raise e
-      end
+  end
     end
   end
 
@@ -73,7 +102,7 @@ module DistributeReads
               Thread.current[:distribute_reads][:primary] = true
             else
               raise DistributeReads::TooMuchLag, "Replica lag over #{max_lag} seconds#{options[:lag_on] ? " on #{base_model.name} connection" : ""}"
-            end
+  end
           end
         end
       end
@@ -83,8 +112,8 @@ module DistributeReads
       value
     ensure
       Thread.current[:distribute_reads] = previous_value
-    end
   end
+end
 end
 
 Makara::Proxy.send :prepend, DistributeReads::AppropriatePool